refactor(core): extract config helpers to app/config.ts
Verified: precommit-verify ✕
